lisätty: tehtävä 2.3 ja 2.4

+++ Ohjelmisto01/moduuli 2/tehtävä02,3.py
@@ -1,5 +1,5 @@
-korkeus = int(input("korkeus:"))
-kanta = int(input("kanta:"))
+korkeus = float(input("korkeus:"))
+kanta = float(input("kanta:"))
 piiri = korkeus*2+kanta*2
 pintaala = korkeus*kanta
 print("suorakulmion piiri :",piiri)
